+++ src/gui/widgets/map/__init__.py
@@ -11,7 +11,7 @@ Sub-components of MapGraphicsView:
 """
 
 from src.gui.widgets.map.drawing_tool import DrawingTool
-from src.gui.widgets.map.icon_picker_dialog import IconPickerDialog
+from src.gui.dialogs.icon_picker_dialog import IconPickerDialog
 from src.gui.widgets.map.interaction_handler import InteractionHandler
 from src.gui.widgets.map.map_graphics_view import MapGraphicsView
 from src.gui.widgets.map.marker_item import MarkerItem